Modify Create#UsersBox so User can add other users
Modify the create method in the UsersBoxesController, so that a User
that owns a Box can give another User access to that Box. Uses
current_user to build initial entry to ensure that user has write access
then changes the user_id of the new entry to whatever is passed in the
params.